UPSTREAM: drm/i915/gt: Correct prefix on a few registers
We have a few registers that have existed for several hardware
generations, but are only used by the driver on Xe_HP and beyond.  In
cases where the Xe_HP version of the register is now replicated and uses
multicast behavior, but earlier generations were singleton, let's change
the register prefix to "XEHP_" to help clarify that we're using the
newer multicast form of the register.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Balasubramani Vivekanandan <balasubramani.vivekanandan@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20221014230239.1023689-5-matthew.d.roper@intel.com
(cherry picked from commit e4abeab94658cdf27f75a824f33ab9ad81d47f96)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:265020278, b:245221509
TEST= 1. This series can improve ~100ms kernel boot time on ADL-N platfrom.
      2. Pass HDMI 2.1 CTS.

Change-Id: I8cd0d8742c91ef210001e77c5c39ad768a4fbc13