UPSTREAM: drm/i915/guc: Handle save/restore of MCR registers explicitly
MCR registers can be placed on the GuC's save/restore list, but at the
moment they are always handled in a multicast manner (i.e., the GuC
reads one instance to save the value and then does a multicast write to
restore that single value to all instances).  In the future the GuC will
probably give us an alternate interface to do unicast per-instance
save/restore operations, so we should be very clear about which
registers on the list are MCR registers (and in the future which
save/restore behavior we want for them).

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Balasubramani Vivekanandan <balasubramani.vivekanandan@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20221014230239.1023689-11-matthew.d.roper@intel.com
(cherry picked from commit cf35f6afb92643633f4ecbb386ab8a572cca0386)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:265020278, b:245221509
TEST= 1. This series can improve ~100ms kernel boot time on ADL-N platfrom.
      2. Pass HDMI 2.1 CTS.

Change-Id: Ife436234df167bc28e73f9aca4fb8bb492ccc698