UPSTREAM: wifi: mt76: mt7921u: fix race issue between reset and suspend/resume
It is unexpected that the reset work is running simultaneously with
the suspend or resume context and it is possible that reset work is still
running even after mt7921 is suspended if we don't fix the race issue.

Thus, the suspend procedure should be waiting until the reset is completed
at the beginning and ignore the subsequent the reset requests.

In case there is an error that happens during either suspend or resume
handler, we will schedule a reset task to recover the error before
returning the error code to ensure we can immediately fix the error there.

Fixes: df3e4143ba8a ("mt76: mt7921u: add suspend/resume support")
Co-developed-by: YN Chen <YN.Chen@mediatek.com>
Signed-off-by: YN Chen <YN.Chen@mediatek.com>
Signed-off-by: Sean Wang <sean.wang@mediatek.com>
Signed-off-by: Felix Fietkau <nbd@nbd.name>
(cherry picked from commit 86f15d043ba7f13211d5c3e41961c3381fb12880)

BUG=b:178754244
TEST=emerge-corsola sys-kernel/chromeos-kernel-5_15;

Signed-off-by: Linux Patches Robot <linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com>
Change-Id: Idfd0af484c41ba27e8046c6eee812c35f8b8b754
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3969471
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Commit-Queue: Abhishek Kumar <kuabhs@chromium.org>
Tested-by: Abhishek Kumar <kuabhs@chromium.org>