UPSTREAM: net: ipa: check table memory regions earlier
Verify that the sizes of the routing and filter table memory regions
are valid as part of memory initialization, rather than waiting for
table initialization.  The main reason to do this is that upcoming
patches use these memory region sizes to determine the number of
entries in these tables, and we'll want to know these sizes are good
sooner.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Paolo Abeni <pabeni@redhat.com>
(cherry picked from commit 73da9cac517ce4dc3a00b87f800846dcc2469ddf)

BUG=b:264290644
TEST=modem boots on trogdor and herobrine

Change-Id: I3f19246080cf1ec5f8942cd27150aa54f1592a60
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4322348
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Douglas Anderson <dianders@chromium.org>