UPSTREAM: drm/amdgpu: use the last IB as gang leader v2
It turned out that not the last IB specified is the gang leader,
but instead the last job allocated.

This is a bit unfortunate and not very intuitive for the CS
interface, so try to fix this.

Signed-off-by: Christian König <christian.koenig@amd.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20221115094206.6181-1-christian.koenig@amd.com
Tested-by: Timur Kristóf <timur.kristof@gmail.com>
Acked-by: Timur Kristóf <timur.kristof@gmail.com>
Reviewed-by: Alex Deucher <alexander.deucher@amd.com>
Fixes: 4624459c84d7 ("drm/amdgpu: add gang submit frontend v6")
(cherry picked from commit eca13f3c67b6ddfcc61fdb9bb1c5f9a7724e2359)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:265055111
TEST=Build test

Backported using forklift.py

Change-Id: I876595d397ca22c2a0ef0a8ef684d738f32dec5c