UPSTREAM: media: videobuf2: revert "get_userptr: buffers are always writable"
Commit 707947247e95 ("media: videobuf2-vmalloc: get_userptr: buffers are
always writable") caused problems in a corner case (passing read-only
shmem memory as a userptr). So revert this patch.

The original problem for which that commit was originally made is
something that I could not reproduce after reverting it. So just go
back to the way it was for many years, and if problems arise in
the future, then another approach should be taken to resolve it.

This patch is based on a patch from Hirokazu.

Fixes: 707947247e95 ("media: videobuf2-vmalloc: get_userptr: buffers are always writable")
Signed-off-by: Hirokazu Honda <hiroh@chromium.org>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Acked-by: Tomasz Figa <tfiga@chromium.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>

(cherry picked from commit e2fc6edd37ba487c64f4dd09f7118b3e45b12d88
 https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git master)

BUG=b:261660224
TEST=video_encode_accelerator_tests

Change-Id: Ib670ce52700268dc35b9fec7c78fcf7628ee4195
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4091383
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>