CHROMIUM: iwl7000: check for all HW errors in debug dump
The hardware, depending on which part fails or times out,
returns 0xA5A5A5A, or 0x5A5A5A5, with the lowest 4 bits
encoding some further reason/status. In iwl_read_prph()
we return 0x5A5A5A5 if getting NIC access fails, but in
the debug code we don't check for both of the possible
values, so we might attempt to dump even while the NIC
isn't responding.

Add a helper to check for both values and use it in the
FW debug code.

BUG=b:269163426
TEST=basic stability and connectivity

Change-Id: I1c297a53787b87e4e2b8f296c041921338573f4d
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4284061
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>