FROMGIT: Revert "drm/i915: Improve on suspend / resume time with VT-d enabled"
This reverts commit 2ef6efa79fecd5e3457b324155d35524d95f2b6b.

Checking the presence if the IRST (Intel Rapid Start Technology)
through the ACPI to decide whether to rebuild or not the GGTT
puts us at the mercy of the boot firmware and we need to
unnecessarily rely on third parties.

Because now we avoid adding scratch pages to the entire GGTT we
don't need this hack anymore.

Signed-off-by: Andi Shyti <andi.shyti@linux.intel.com>
Cc: Thomas Hellström <thomas.hellstrom@linux.intel.com>
Cc: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20221130235805.221010-6-andi.shyti@linux.intel.com
(cherry picked from commit de3a9ab97069488b23427726e86b8628f4fe278e
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-gt-next)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:265020278, b:245221509
TEST= 1. This series can improve ~100ms kernel boot time on ADL-N platfrom.
      2. Pass HDMI 2.1 CTS.

Change-Id: I6ececeaf38c5679aaf167fb571b28ed59c8cf2af