BACKPORT:FROMLIST:v4l2-ctrls: add support for V4L2_CTRL_WHICH_MIN/MAX_VAL
Add the capability of retrieving the min and max values of a
compound control.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Yunke Cao <yunkec@google.com>
(am from https://patchwork.kernel.org/project/linux-media/patch/20230426082923.132909-7-yunkec@google.com/)

The backport did not the changes in
/drivers/media/v4l2-core/v4l2-ctrls-api.c
/include/media/v4l2-ctrls.h
because these files had a big refactor.
These changes are not required for the UVC driver.

UPSTREAM-TASK=b:191930245
BUG=b:191930245
TEST=v4l2-compliance

Change-Id: If703f3372aadba2ca2be3655483fd45e2a2ba6b0
Signed-off-by: Yunke Cao <yunkec@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4663715
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>